@@ -353,7 +353,7 @@ def get_footprint(tod, wcs_kernel, dets=None, timestamps=None, boresight=None,
 
 
 def get_focal_plane_cover(tod=None, count=0, focal_plane=None,
-                          xieta=None):
+                          xieta=None, det_weights=None):
     """Process a bunch of detector positions into a center and radius such
     that a circle with that center and radius contains all the
     detectors.  Also return detector positions, arranged approximately
@@ -366,6 +366,9 @@ def get_focal_plane_cover(tod=None, count=0, focal_plane=None,
         not passed in
       xieta (array): (2, n) array (or similar) of xi and eta
         detector positions.
+      det_weights (array): If provided, must be same length as the xi
+        and eta vectors. Only dets with non-zero value for det_weights
+        will be included in the evaluation of the cover.
 
     Returns:
       xieta0: array[2] with array center, (xi0, eta0).
@@ -374,8 +377,21 @@ def get_focal_plane_cover(tod=None, count=0, focal_plane=None,
         coords of the circular convex hull.
 
     Notes:
-      If count=0, an empty list is returned for xietas.  Otherwise,
+      If count=0, an empty list is returned for xietas. Otherwise,
       count must be at least 3 so that the shape is not degenerate.
+
+      Any xi, eta that are not finite (e.g. nan or inf) are excluded
+      from the computation. If no detectors remain after the combined
+      finiteness and det_weights cuts, a ValueError is raised.
+
+      Note that ``det_weights`` can be int, float, or bool
+      type. Sometimes you might want to only include optical dets in
+      the result; e.g.::
+
+          ..., det_weights=(aman.det_info.wafer.type=="OPTC"), ...
+
+      In degenerate cases (all dets are in exactly the same place), a
+      radius of zero may be returned.
 
     """
     if xieta is None:
@@ -385,6 +401,17 @@ def get_focal_plane_cover(tod=None, count=0, focal_plane=None,
         eta = focal_plane.eta
     else:
         xi, eta = xieta[:2]
+    mask = np.isfinite(xi) * np.isfinite(eta)
+    if det_weights is not None:
+        mask *= det_weights.astype(bool)
+
+    if not np.any(mask):
+        raise ValueError('All provided (xi, eta) coords are excluded; '
+                         'cannot estimate a focal plane cover.')
+
+    # Restrict to only dets under consideration.
+    xi, eta = xi[mask], eta[mask]
+
     qs = so3g.proj.quat.rotation_xieta(xi, eta)
 
     # Starting guess for center
@@ -477,23 +504,23 @@ def get_supergeom(*geoms, tol=1e-3):
         s0 = corner_b - corner_a
     return tuple(map(int, s0)), w0
 
-def _confirm_wcs(*maps):
-    """Insist that all arguments either have the same .wcs, or do not have
-    a wcs.  Each argument should be either an ndmap (with a .wcs
-    attribute) or an ndarray (without a .wcs attribute).
+def _confirm_wcs(*wcss):
+    """Insist that all arguments are either the same wcs, or None.
 
-    Raises a ValueError if more than one argument has a .wcs attribute
-    and they do not all agree.  Returns either the first .wcs
-    attribute encountered, or None if there aren't any.
+    Raises a ValueError if more than one argument is a wcs, and not
+    all wcs agree.  Returns either the first valid wcs, or None if
+    there aren't any.
 
     """
     wcs_to_use = None
-    for i, m in enumerate(maps):
-        if hasattr(m, 'wcs'):
-            if wcs_to_use is None:
-                wcs_to_use = m.wcs
-            elif not wcsutils.equal(wcs_to_use, m.wcs):
-                raise ValueError('The wcs from %ith item is discordant with prior ones.' % i)
+    for i, wcs in enumerate(wcss):
+        if wcs is None:
+            continue
+        if wcs_to_use is None:
+            wcs_to_use = wcs
+        elif not wcsutils.equal(wcs_to_use, wcs):
+            raise ValueError(
+                f'The wcs from {i}th item ({wcs}) is discordant with prior ones ({wcs_to_use})')
     return wcs_to_use
 
 def _invert_weights_map(weights, eigentol=1e-6, kill_partials=True,
@@ -565,23 +592,23 @@ def _invert_weights_map(weights, eigentol=1e-6, kill_partials=True,
     # Reshape the output to match what was passed in.
     return iw.transpose(1,2,0).reshape(weights.shape)
 
-def _apply_inverse_weights_map(inverse_weights, target):
+def _apply_inverse_weights_map(inverse_weights, target, out=None):
     """Apply a map of matrices to a map of vectors.
 
     Assumes inverse_weights.shape = (a, b, ...) and target.shape =
     (b, ...); the result has shape (a, ...).
 
     """
-    # master had:
-    #iw = inverse_weights.transpose((2,3,0,1))
-    #m = target.transpose((1,2,0)).reshape(
-    #    target.shape[1], target.shape[2], target.shape[0], 1)
-    #m1 = np.matmul(iw, m)
-    #return m1.transpose(2,3,0,1).reshape(target.shape)
+    if out is None:
+        out = np.empty(inverse_weights.shape[1:],
+                       dtype=target.dtype)
+    # Recall matmul(a, b) operates on the last two axes of (a, b). So
+    # move axes, and create a second one in target; re-order at end.
     iw = np.moveaxis(inverse_weights, (0,1), (-2,-1))
     t  = np.moveaxis(target[:,None],  (0,1), (-2,-1))
-    m  = np.matmul(iw, t)
-    return np.moveaxis(m, (-2,-1), (0,1))[:,0]
+    out_moved = np.moveaxis(out[:,None], (0,1), (-2,-1))
+    np.matmul(iw, t, out=out_moved)
+    return out
 
 class ScalarLastQuat(np.ndarray):
     """Wrapper class for numpy arrays carrying quaternions with the ijk1
