@@ -182,7 +182,7 @@ def from_map(tod, signal_map, cuts=None, flip_gamma=True, wrap=False, modulated=
         return signal_sim
 
 
-def rotate_demodQU(tod, update_focal_plane=True):
+def rotate_demodQU(tod, sign=1, offset=0, update_focal_plane=True):
     """
     Apply detectors' polarization angle calibration to the HWP demodulated Q and U timestreams to
     place all detectors' Q and U timestreams in a common telescope frame. This updates tod.demodQ
@@ -190,12 +190,16 @@ def rotate_demodQU(tod, update_focal_plane=True):
 
     Args:
         tod : an axisManager object
-            update_focal_plane (bool, optional): Whether to set focal_plane.gamma angles to zero,
+        update_focal_plane (bool, optional): Whether to set focal_plane.gamma angles to zero,
             consistent with new coordinate reference. Make this true for polarization mapmaking
             using make_map.
+        offset : float, optional
+            The rotation angle in degrees to apply (default is 0).
+        sign : int, optional
+            A sign factor to control the direction of the rotation (default is +1).
 
     """
-    demodC = ((tod.demodQ + 1j*tod.demodU).T * np.exp(-2j*tod.focal_plane.gamma)).T
+    demodC = ((tod.demodQ + 1j*tod.demodU).T * np.exp( sign*(-2j*tod.focal_plane.gamma + 1j*np.deg2rad(offset)) )).T
     tod.demodQ = demodC.real
     tod.demodU = demodC.imag
     del demodC
