@@ -29,33 +29,26 @@ def dummy_preproc(obs_id, group_list, logger,
     error = None
     outputs = []
     context = core.Context(configs["context_file"])
-    group_by, groups = pp_util.get_groups(obs_id, configs, context)
+    group_by, groups, error = pp_util.get_groups(obs_id, configs, context)
     pipe = Pipeline(configs["process_pipe"], plot_dir=configs["plot_dir"], logger=logger)
     for group in groups:
         logger.info(f"Beginning run for {obs_id}:{group}")
+        dets = {gb:gg for gb, gg in zip(group_by, group)}
         proc_aman = core.AxisManager(core.LabelAxis('dets', ['det%i' % i for i in range(3)]),
                                      core.OffsetAxis('samps', 1000))
         proc_aman.wrap_new('signal', ('dets', 'samps'), dtype='float32')
         proc_aman.wrap_new('timestamps', ('samps',))[:] = (np.arange(proc_aman.samps.count) / 200)
-        policy = pp_util.ArchivePolicy.from_params(configs['archive']['policy'])
-        dest_file, dest_dataset = policy.get_dest(obs_id)
-        for gb, g in zip(group_by, group):
-            if gb == 'detset':
-                dest_dataset += "_" + g
-            else:
-                dest_dataset += "_" + gb + "_" + str(g)
-        logger.info(f"Saving data to {dest_file}:{dest_dataset}")
-        proc_aman.save(dest_file, dest_dataset, overwrite)
 
-        # Collect index info.
-        db_data = {'obs:obs_id': obs_id,
-                   'dataset': dest_dataset}
-        for gb, g in zip(group_by, group):
-            db_data['dets:'+gb] = g
+        outputs_grp = pp_util.save_group(obs_id, configs, dets, context, subdir='temp')
+        logger.info(f"Saving data to {outputs_grp['temp_file']}:{outputs_grp['db_data']['dataset']}")
+        proc_aman.save(outputs_grp['temp_file'], outputs_grp['db_data']['dataset'], overwrite)
+
         if run_parallel:
-            outputs.append(db_data)
+            outputs.append(outputs_grp)
+
     if run_parallel:
-        return error, dest_file, outputs
+        return error, outputs
+
 
 def preprocess_tod(obs_id,
                    configs,
@@ -90,7 +83,14 @@ def preprocess_tod(obs_id,
         configs = yaml.safe_load(open(configs, "r"))
 
     context = core.Context(configs["context_file"])
-    group_by, groups = pp_util.get_groups(obs_id, configs, context)
+    group_by, groups, error = pp_util.get_groups(obs_id, configs, context)
+
+    if error is not None:
+        if run_parallel:
+            return error[0], [None, None]
+        else:
+            return
+
     all_groups = groups.copy()
     for g in all_groups:
         if group_list is not None:
@@ -119,7 +119,7 @@ def preprocess_tod(obs_id,
                        f"No analysis to run.")
         error = 'no_group_overlap'
         if run_parallel:
-            return error, None, [None, None]
+            return error, [None, None]
         else:
             return
 
@@ -136,8 +136,9 @@ def preprocess_tod(obs_id,
     n_fail = 0
     for group in groups:
         logger.info(f"Beginning run for {obs_id}:{group}")
+        dets = {gb:gg for gb, gg in zip(group_by, group)}
         try:
-            aman = context.get_obs(obs_id, dets={gb:g for gb, g in zip(group_by, group)})
+            aman = context.get_obs(obs_id, dets=dets)
             tags = np.array(context.obsdb.get(aman.obs_info.obs_id, tags=True)['tags'])
             aman.wrap('tags', tags)
             proc_aman, success = pipe.run(aman)
@@ -161,29 +162,18 @@ def preprocess_tod(obs_id,
             n_fail += 1
             continue
 
-        policy = pp_util.ArchivePolicy.from_params(configs['archive']['policy'])
-        dest_file, dest_dataset = policy.get_dest(obs_id)
-        for gb, g in zip(group_by, group):
-            if gb == 'detset':
-                dest_dataset += "_" + g
-            else:
-                dest_dataset += "_" + gb + "_" + str(g)
-        logger.info(f"Saving data to {dest_file}:{dest_dataset}")
-        proc_aman.save(dest_file, dest_dataset, overwrite)
+        outputs_grp = pp_util.save_group(obs_id, configs, dets, context, subdir='temp')
+        logger.info(f"Saving data to {outputs_grp['temp_file']}:{outputs_grp['db_data']['dataset']}")
+        proc_aman.save(outputs_grp['temp_file'], outputs_grp['db_data']['dataset'], overwrite)
 
-        # Collect index info.
-        db_data = {'obs:obs_id': obs_id,
-                'dataset': dest_dataset}
-        for gb, g in zip(group_by, group):
-            db_data['dets:'+gb] = g
         if run_parallel:
-            outputs.append(db_data)
+            outputs.append(outputs_grp)
         else:
-            logger.info(f"Saving to database under {db_data}")
-            if len(db.inspect(db_data)) == 0:
-                h5_path = os.path.relpath(dest_file,
+            logger.info(f"Saving to database under {outputs_grp['db_data']}")
+            if len(db.inspect(outputs_grp['db_data'])) == 0:
+                h5_path = os.path.relpath(outputs_grp['temp_file'],
                         start=os.path.dirname(configs['archive']['index']))
-                db.add_entry(db_data, h5_path)
+                db.add_entry(outputs_grp['db_data'], h5_path)
 
     if make_lmsi:
         from pathlib import Path
@@ -199,17 +189,18 @@ def preprocess_tod(obs_id,
             # If no groups make it to the end of the processing return error.
             logger.info(f'ERROR: all groups failed for {obs_id}')
             error = 'all_fail'
-            return error, None, [obs_id, 'all groups']
+            return error, [obs_id, 'all groups']
         else:
             logger.info('Returning data to futures')
             error = None
-            return error, dest_file, outputs
+            return error, outputs
+
 
 def load_preprocess_tod_sim(obs_id, sim_map,
                             configs="preprocess_configs.yaml",
                             context=None, dets=None,
                             meta=None, modulated=True):
-    """ Loads the saved information from the preprocessing pipeline and runs the
+    """Loads the saved information from the preprocessing pipeline and runs the
     processing section of the pipeline on simulated data
 
     Assumes preprocess_tod has already been run on the requested observation.
@@ -235,7 +226,8 @@ def load_preprocess_tod_sim(obs_id, sim_map,
     """
     configs, context = pp_util.get_preprocess_context(configs, context)
     meta = pp_util.load_preprocess_det_select(obs_id, configs=configs,
-                                              context=context, dets=dets, meta=meta)
+                                              context=context, dets=dets,
+                                              meta=meta)
 
     if meta.dets.count == 0:
         logger.info(f"No detectors left after cuts in obs {obs_id}")
@@ -252,6 +244,7 @@ def load_preprocess_tod_sim(obs_id, sim_map,
         demod_mm.from_map(aman, sim_map, wrap=True, modulated=modulated)
         pipe.run(aman, aman.preprocess, sim=True)
         return aman
+
 
 def get_parser(parser=None):
     if parser is None:
@@ -310,6 +303,7 @@ def get_parser(parser=None):
     )
     return parser
 
+
 def main(
         configs: str,
         query: Optional[str] = None,
@@ -333,8 +327,17 @@ def main(
     obs_list = sp_util.get_obslist(context, query=query, obs_id=obs_id, min_ctime=min_ctime,
                                    max_ctime=max_ctime, update_delay=update_delay, tags=tags,
                                    planet_obs=planet_obs)
+
     if len(obs_list)==0:
         logger.warning(f"No observations returned from query: {query}")
+
+    # clean up lingering files from previous incomplete runs
+    policy_dir = os.path.join(os.path.dirname(configs['archive']['policy']['filename']), 'temp')
+    for obs in obs_list:
+        obs_id = obs['obs_id']
+        pp_util.cleanup_obs(obs_id, policy_dir, errlog, configs, context,
+                            subdir='temp', remove=overwrite)
+
     run_list = []
 
     if overwrite or not os.path.exists(configs['archive']['index']):
@@ -345,40 +348,26 @@ def main(
         db = core.metadata.ManifestDb(configs['archive']['index'])
         for obs in obs_list:
             x = db.inspect({'obs:obs_id': obs["obs_id"]})
-            group_by, groups = pp_util.get_groups(obs["obs_id"], configs, context)
             if x is None or len(x) == 0:
                 run_list.append( (obs, None) )
-            elif len(x) != len(groups):
-                [groups.remove([a[f'dets:{gb}'] for gb in group_by]) for a in x]
-                run_list.append( (obs, groups) )
+            else:
+                group_by, groups, _ = pp_util.get_groups(obs["obs_id"], configs, context)
+                if len(x) != len(groups):
+                    [groups.remove([a[f'dets:{gb}'] for gb in group_by]) for a in x]
+                    run_list.append( (obs, groups) )
 
     logger.info(f'Run list created with {len(run_list)} obsids')
-
-    # Expects archive policy filename to be <path>/<filename>.h5 and then this adds
-    # <path>/<filename>_<xxx>.h5 where xxx is a number that increments up from 0 
-    # whenever the file size exceeds 10 GB.
-    nfile = 0
-    folder = os.path.dirname(configs['archive']['policy']['filename'])
-    basename = os.path.splitext(configs['archive']['policy']['filename'])[0]
-    dest_file = basename + '_' + str(nfile).zfill(3) + '.h5'
-    if not(os.path.exists(folder)):
-            os.makedirs(folder)
-    while os.path.exists(dest_file) and os.path.getsize(dest_file) > 10e9:
-        nfile += 1
-        dest_file = basename + '_' + str(nfile).zfill(3) + '.h5'
-
-    logger.info(f'Starting dest_file set to {dest_file}')
 
     # Run write_block obs-ids in parallel at once then write all to the sqlite db.
     with ProcessPoolExecutor(nproc) as exe:
         futures = [exe.submit(preprocess_tod, obs_id=r[0]['obs_id'],
                      group_list=r[1], verbosity=verbosity,
-                     configs=pp_util.swap_archive(configs, f'temp/{r[0]["obs_id"]}.h5'),
+                     configs=configs,
                      overwrite=overwrite, run_parallel=True) for r in run_list]
         for future in as_completed(futures):
             logger.info('New future as_completed result')
             try:
-                err, src_file, db_datasets = future.result()
+                err, db_datasets = future.result()
             except Exception as e:
                 errmsg = f'{type(e)}: {e}'
                 tb = ''.join(traceback.format_tb(e.__traceback__))
@@ -389,37 +378,10 @@ def main(
                 continue
             futures.remove(future)
 
-            logger.info(f'Processing future result db_dataset: {db_datasets}')
-            db = pp_util.get_preprocess_db(configs, group_by)
-            logger.info('Database connected')
-            if os.path.exists(dest_file) and os.path.getsize(dest_file) >= 10e9:
-                nfile += 1
-                dest_file = basename + '_'+str(nfile).zfill(3)+'.h5'
-                logger.info('Starting a new h5 file.')
-
-            h5_path = os.path.relpath(dest_file,
-                            start=os.path.dirname(configs['archive']['index']))
-
-            if err is None:
-                logger.info(f'Moving files from temp to final destination.')
-                with h5py.File(dest_file,'a') as f_dest:
-                    with h5py.File(src_file,'r') as f_src:
-                        for dts in f_src.keys():
-                            f_src.copy(f_src[f'{dts}'], f_dest, f'{dts}')
-                            for member in f_src[dts]:
-                                if isinstance(f_src[f'{dts}/{member}'], h5py.Dataset):
-                                    f_src.copy(f_src[f'{dts}/{member}'], f_dest[f'{dts}'], f'{dts}/{member}')
-                for db_data in db_datasets:
-                    logger.info(f"Saving to database under {db_data}")
-                    if len(db.inspect(db_data)) == 0:
-                        db.add_entry(db_data, h5_path)
-                logger.info(f'Deleting {src_file}.')
-                os.remove(src_file)
-            else:
-                logger.info(f'Writing {db_datasets[0]} to error log')
-                f = open(errlog, 'a')
-                f.write(f'\n{time.time()}, {err}, {db_datasets[0]}\n{db_datasets[1]}\n')
-                f.close()
+            if db_datasets:
+                logger.info(f'Processing future result db_dataset: {db_datasets}')
+                for db_dataset in db_datasets:
+                    pp_util.cleanup_mandb(err, db_dataset, configs, logger)
 
 if __name__ == '__main__':
     sp_util.main_launcher(main, get_parser)
