@@ -213,7 +213,8 @@ def calc_psd(
     prefer='center',
     freq_spacing=None,
     merge=False, 
-    overwrite=True, 
+    overwrite=True,
+    subscan=False,
     **kwargs
 ):
     """Calculates the power spectrum density of an input signal using signal.welch().
@@ -234,6 +235,7 @@ def calc_psd(
             If an nperseg is explicitly passed then that will be used.
         merge (bool): if True merge results into axismanager.
         overwrite (bool): if true will overwrite f, pxx axes.
+        subscan (bool): if True, compute psd on subscans.
         **kwargs: keyword args to be passed to signal.welch().
 
     Returns:
@@ -242,34 +244,40 @@ def calc_psd(
     """
     if signal is None:
         signal = aman.signal
-    if timestamps is None:
-        timestamps = aman.timestamps
-
-    n_samps = signal.shape[-1]
-    if n_samps <= max_samples:
-        start = 0
-        stop = n_samps
+    if subscan:
+        freqs, Pxx = _calc_psd_subscan(aman, signal=signal, freq_spacing=freq_spacing, **kwargs)
+        axis_map_pxx = [(0, "dets"), (1, "nusamps"), (2, "subscans")]
     else:
-        offset = n_samps - max_samples
-        if prefer == "left":
-            offset = 0
-        elif prefer == "center":
-            offset //= 2
-        elif prefer == "right":
-            pass
-        else:
-            raise ValueError(f"Invalid choise prefer='{prefer}'")
-        start = offset
-        stop = offset + max_samples
-    fs = 1 / np.nanmedian(np.diff(timestamps[start:stop]))
-    if "nperseg" not in kwargs:
-        if freq_spacing is not None:
-            nperseg = int(2 ** (np.around(np.log2(fs / freq_spacing))))
-        else:
-            nperseg = int(2 ** (np.around(np.log2((stop - start) / 50.0))))
-        kwargs["nperseg"] = nperseg
+        if timestamps is None:
+            timestamps = aman.timestamps
 
-    freqs, Pxx = welch(signal[:, start:stop], fs, **kwargs)
+        n_samps = signal.shape[-1]
+        if n_samps <= max_samples:
+            start = 0
+            stop = n_samps
+        else:
+            offset = n_samps - max_samples
+            if prefer == "left":
+                offset = 0
+            elif prefer == "center":
+                offset //= 2
+            elif prefer == "right":
+                pass
+            else:
+                raise ValueError(f"Invalid choice prefer='{prefer}'")
+            start = offset
+            stop = offset + max_samples
+        fs = 1 / np.nanmedian(np.diff(timestamps[start:stop]))
+        if "nperseg" not in kwargs:
+            if freq_spacing is not None:
+                nperseg = int(2 ** (np.around(np.log2(fs / freq_spacing))))
+            else:
+                nperseg = int(2 ** (np.around(np.log2((stop - start) / 50.0))))
+            kwargs["nperseg"] = nperseg
+
+        freqs, Pxx = welch(signal[:, start:stop], fs, **kwargs)
+        axis_map_pxx = [(0, aman.dets), (1, "nusamps")]
+
     if merge:
         aman.merge( core.AxisManager(core.OffsetAxis("nusamps", len(freqs))))
         if overwrite:
@@ -278,9 +286,42 @@ def calc_psd(
             if "Pxx" in aman._fields:
                 aman.move("Pxx", None)
         aman.wrap("freqs", freqs, [(0,"nusamps")])
-        aman.wrap("Pxx", Pxx, [(0,"dets"),(1,"nusamps")])
+        aman.wrap("Pxx", Pxx, axis_map_pxx)
     return freqs, Pxx
 
+def _calc_psd_subscan(aman, signal=None, freq_spacing=None, **kwargs):
+    """
+    Calculate the power spectrum density of subscans using signal.welch().
+    Data defaults to aman.signal. aman.timestamps is used for times.
+    aman.subscan_info is used to identify subscans.
+    See calc_psd for arguments.
+    """
+    from .flags import get_subscan_signal
+    if signal is None:
+        signal = aman.signal
+
+    fs = 1 / np.nanmedian(np.diff(aman.timestamps))
+    if "nperseg" not in kwargs:
+        if freq_spacing is not None:
+            nperseg = int(2 ** (np.around(np.log2(fs / freq_spacing))))
+        else:
+            duration_samps = np.asarray([np.ptp(x.ranges()) if x.ranges().size > 0 else 0 for x in aman.subscan_info.subscan_flags])
+            duration_samps = duration_samps[duration_samps > 0]
+            nperseg = int(2 ** (np.around(np.log2(np.median(duration_samps) / 4))))
+        kwargs["nperseg"] = nperseg
+
+    Pxx = []
+    for iss in range(aman.subscan_info.subscans.count):
+        signal_ss = get_subscan_signal(aman, signal, iss)
+        axis = -1 if "axis" not in kwargs else kwargs["axis"]
+        if signal_ss.shape[axis] >= kwargs["nperseg"]:
+            freqs, pxx_sub = welch(signal_ss, fs, **kwargs)
+            Pxx.append(pxx_sub)
+        else:
+            Pxx.append(np.full((signal.shape[0], kwargs["nperseg"]//2+1), np.nan)) # Add nans if subscan is too short
+    Pxx = np.array(Pxx)
+    Pxx = Pxx.transpose(1, 2, 0) # Dets, nusamps, subscans
+    return freqs, Pxx
 
 def calc_wn(aman, pxx=None, freqs=None, low_f=5, high_f=10):
     """
@@ -346,13 +387,15 @@ def fit_noise_model(
     signal=None,
     f=None,
     pxx=None,
-    psdargs=None,
+    psdargs={},
     fwhite=(10, 100),
     lowf=1,
     merge_fit=False,
     f_max=100,
     merge_name="noise_fit_stats",
     merge_psd=True,
+    freq_spacing=None,
+    subscan=False
 ):
     """
     Fits noise model with white and 1/f noise to the PSD of signal.
@@ -392,6 +435,10 @@ def fit_noise_model(
         If ``merge_fit`` is True then addes into axis manager with merge_name.
     merge_psd : bool
         If ``merg_psd`` is True then adds fres and Pxx to the axis manager.
+    freq_spacing : float
+        The approximate desired frequency spacing of the PSD. Passed to calc_psd.
+    subscan : bool
+        If True, fit noise on subscans.
     Returns
     -------
     noise_fit_stats : AxisManager
@@ -403,42 +450,48 @@ def fit_noise_model(
         signal = aman.signal
 
     if f is None or pxx is None:
-        if psdargs is None:
-            f, pxx = calc_psd(
-                aman, signal=signal, timestamps=aman.timestamps, merge=merge_psd
-            )
-        else:
-            f, pxx = calc_psd(
-                aman,
-                signal=signal,
-                timestamps=aman.timestamps,
-                merge=merge_psd,
-                **psdargs,
-            )
-    eix = np.argmin(np.abs(f - f_max))
-    f = f[1:eix]
-    pxx = pxx[:, 1:eix]
+        f, pxx = calc_psd(
+            aman,
+            signal=signal,
+            timestamps=aman.timestamps,
+            freq_spacing=freq_spacing,
+            merge=merge_psd,
+            subscan=subscan,
+            **psdargs,
+        )
+    if subscan:
+       fitout, covout = _fit_noise_model_subscan(aman, signal,  f, pxx, psdargs=psdargs,
+                                                  fwhite=fwhite, lowf=lowf, f_max=f_max,
+                                                  freq_spacing=freq_spacing)
+       axis_map_fit = [(0, "dets"), (1, "noise_model_coeffs"), (2, aman.subscans)]
+       axis_map_cov = [(0, "dets"), (1, "noise_model_coeffs"), (2, "noise_model_coeffs"), (3, aman.subscans)]
+    else:
+        eix = np.argmin(np.abs(f - f_max))
+        f = f[1:eix]
+        pxx = pxx[:, 1:eix]
 
-    fitout = np.zeros((aman.dets.count, 3))
-    # This is equal to np.sqrt(np.diag(cov)) when doing curve_fit
-    covout = np.zeros((aman.dets.count, 3, 3))
-    for i in range(aman.dets.count):
-        p = pxx[i]
-        wnest = np.median(p[((f > fwhite[0]) & (f < fwhite[1]))])
-        pfit = np.polyfit(np.log10(f[f < lowf]), np.log10(p[f < lowf]), 1)
-        fidx = np.argmin(np.abs(10 ** np.polyval(pfit, np.log10(f)) - wnest))
-        p0 = [f[fidx], wnest, -pfit[0]]
-        bounds = [(0, None), (sys.float_info.min, None), (None, None)]
-        res = minimize(neglnlike, p0, args=(f, p), bounds=bounds, method="Nelder-Mead")
-        try:
-            Hfun = ndt.Hessian(lambda params: neglnlike(params, f, p), full_output=True)
-            hessian_ndt, _ = Hfun(res["x"])
-            # Inverse of the hessian is an estimator of the covariance matrix
-            # sqrt of the diagonals gives you the standard errors.
-            covout[i] = np.linalg.inv(hessian_ndt)
-        except np.linalg.LinAlgError:
-            covout[i] = np.full((3, 3), np.nan)
-        fitout[i] = res.x
+        fitout = np.zeros((aman.dets.count, 3))
+        # This is equal to np.sqrt(np.diag(cov)) when doing curve_fit
+        covout = np.zeros((aman.dets.count, 3, 3))
+        for i in range(aman.dets.count):
+            p = pxx[i]
+            wnest = np.median(p[((f > fwhite[0]) & (f < fwhite[1]))])
+            pfit = np.polyfit(np.log10(f[f < lowf]), np.log10(p[f < lowf]), 1)
+            fidx = np.argmin(np.abs(10 ** np.polyval(pfit, np.log10(f)) - wnest))
+            p0 = [f[fidx], wnest, -pfit[0]]
+            bounds = [(0, None), (sys.float_info.min, None), (None, None)]
+            res = minimize(neglnlike, p0, args=(f, p), bounds=bounds, method="Nelder-Mead")
+            try:
+                Hfun = ndt.Hessian(lambda params: neglnlike(params, f, p), full_output=True)
+                hessian_ndt, _ = Hfun(res["x"])
+                # Inverse of the hessian is an estimator of the covariance matrix
+                # sqrt of the diagonals gives you the standard errors.
+                covout[i] = np.linalg.inv(hessian_ndt)
+            except np.linalg.LinAlgError:
+                covout[i] = np.full((3, 3), np.nan)
+            fitout[i] = res.x
+        axis_map_fit = [(0, "dets"), (1, "noise_model_coeffs")]
+        axis_map_cov = [(0, "dets"), (1, "noise_model_coeffs"), (2, "noise_model_coeffs")]
 
     noise_model_coeffs = ["fknee", "white_noise", "alpha"]
     noise_fit_stats = core.AxisManager(
@@ -447,16 +500,44 @@ def fit_noise_model(
             name="noise_model_coeffs", vals=np.array(noise_model_coeffs, dtype="<U8")
         ),
     )
-    noise_fit_stats.wrap("fit", fitout, [(0, "dets"), (1, "noise_model_coeffs")])
-    noise_fit_stats.wrap(
-        "cov",
-        covout,
-        [(0, "dets"), (1, "noise_model_coeffs"), (2, "noise_model_coeffs")],
-    )
+    noise_fit_stats.wrap("fit", fitout, axis_map_fit)
+    noise_fit_stats.wrap("cov", covout, axis_map_cov)
 
     if merge_fit:
         aman.wrap(merge_name, noise_fit_stats)
     return noise_fit_stats
+
+
+def _fit_noise_model_subscan(
+    aman,
+    signal,
+    f,
+    pxx,
+    psdargs={},
+    fwhite=(10, 100),
+    lowf=1,
+    f_max=100,
+    freq_spacing=None,
+):
+    """
+    Fits noise model with white and 1/f noise to the PSD of signal subscans.
+    Args are as for fit_noise_model.
+    """
+    fitout = np.empty((aman.dets.count, 3, aman.subscan_info.subscans.count))
+    covout = np.empty((aman.dets.count, 3, 3, aman.subscan_info.subscans.count))
+
+    for isub in range(aman.subscan_info.subscans.count):
+        if np.all(np.isnan(pxx[...,isub])): # Subscan has been fully cut
+            fitout[..., isub] = np.full((aman.dets.count, 3), np.nan)
+            covout[..., isub] = np.full((aman.dets.count, 3, 3), np.nan)
+        else:
+            noise_model = fit_noise_model(aman, f=f, pxx=pxx[...,isub], fwhite=fwhite, lowf=lowf, merge_fit=False,
+                                          f_max=f_max, merge_psd=False, subscan=False)
+
+            fitout[..., isub] = noise_model.fit
+            covout[..., isub] = noise_model.cov
+
+    return fitout, covout
 
 
 def build_hpf_params_dict(
